:recycle: Split Connection class into multiple files
